[MIST-159] Refactor TextSocketSourceGenerator
